+++ src/assets/jss/material-dashboard-react/components/snackbarContentStyle.jsx
@@ -30,6 +30,8 @@ const snackbarContentStyle = {
     backgroundColor: whiteColor,
     color: grayColor[7],
     borderRadius: "3px",
+    minWidth: "unset",
+    maxWidth: "unset",
     boxShadow:
       "0 12px 20px -10px rgba(" +
       hexToRgb(whiteColor) +
@@ -119,6 +121,10 @@ const snackbarContentStyle = {
   iconMessage: {
     paddingLeft: "50px",
     display: "block"
+  },
+  actionRTL: {
+    marginLeft: "-8px",
+    marginRight: "auto"
   }
 };